requirements.txt
google-cloud-storage
boto3